requirements.txt
anthropic>=0.25.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
APScheduler>=3.10.0
PyYAML>=6.0
pytz>=2024.1
